user/requirements.txt
nats-py==2.2.0
pymongo==4.2.0
mongoengine==0.24.2
bcrypt==4.0.1
pyjwt==2.5.0